requirements.txt
# Wazuh Real-time Data Fetcher Requirements

# Docker integration
docker==7.1.0

# Database
sqlite3  # Built-in Python module

# Data processing
pandas==2.2.2

# Logging
colorlog==6.8.0

# JSON processing (built-in)
# datetime (built-in)
# threading (built-in)

# Additional utilities for monitoring
psutil==6.0.0

# LangChain for RAG
langchain==0.2.16
langchain-community==0.2.12  
langchain-openai==0.1.23
chromadb==0.5.5
sentence-transformers==3.0.1
faiss-cpu==1.8.0

# Development/Testing (optional)
pytest==8.3.3
pytest-asyncio==0.24.0
